Compile warning fix in hash_table_test
Reviewed-on: http://review.northscale.com/2657
Tested-by: Dustin Sallings <dustin@spy.net>

+++ t/hash_table_test.cc
@@ -22,8 +22,8 @@ EPStats global_stats;
 class Counter : public HashTableVisitor {
 public:
 
-    int count;
-    int deleted;
+    size_t count;
+    size_t deleted;
 
     Counter(bool v) : count(), deleted(), verify(v) {}
